feat: trying to push

+++ src/main/java/com/example/naturelink/Service/ReservationService.java
@@ -52,7 +52,7 @@ public class ReservationService implements IReservationService {
             reservation.setTyperes(TypeReservation.LOGEMENT);
             reservation.setNumRooms(dto.getNumRooms()); // Set number of rooms if the reservation is for logement
         } else if (dto.getEventId() != null) {
-            Event event = evenementRepository.findById(Math.toIntExact(dto.getEventId()))
+            Event event = evenementRepository.findById((long) Math.toIntExact(dto.getEventId()))
                     .orElseThrow(() -> new RuntimeException("Event not found"));
             reservation.setEventId(event);
             reservation.setTyperes(TypeReservation.EVENT);
@@ -62,7 +62,7 @@ public class ReservationService implements IReservationService {
             reservation.setRestaurant(restaurant);
             reservation.setTyperes(TypeReservation.RESTAURANT);
         } else if (dto.getTransportId() != null) {
-            Transport transport = transportRepository.findById(Math.toIntExact(Long.valueOf(dto.getTransportId())))
+            Transport transport = transportRepository.findById((long) Math.toIntExact(Long.valueOf(dto.getTransportId())))
                     .orElseThrow(() -> new RuntimeException("Transport not found"));
             reservation.setTranspId(transport);
             reservation.setTyperes(TypeReservation.TRANSPORT);
